Move newLogFields to logrusx
